+++ src/Libraries/SmartStore.Core/Data/Hooks/HookedEntity.cs
@@ -1,18 +1,52 @@
-﻿using System.Data.Entity.Infrastructure;
+﻿using System;
+using System.Data.Entity.Infrastructure;
 
 namespace SmartStore.Core.Data.Hooks
 {
-	public sealed class HookedEntity
+	public interface IHookedEntity
 	{
+		/// <summary>
+		/// Gets the hooked entity entry
+		/// </summary>
+		DbEntityEntry Entry { get; }
+
+		BaseEntity Entity { get; }
+
+		Type EntityType { get; }
+
+		/// <summary>
+		/// Gets or sets the initial (presave) state of the hooked entity.
+		/// The setter is for internal use only, don't invoke!
+		/// </summary>
+		EntityState InitialState { get; set; }
+
+		/// <summary>
+		/// Gets or sets the current state of the hooked entity
+		/// </summary>
+		EntityState State { get; set; }
+
+		/// <summary>
+		/// Gets a value indicating whether the entity state has changed during hooking.
+		/// </summary>
+		bool HasStateChanged { get; }
+
+		/// <summary>
+		/// Gets a value indicating whether a property has been modified.
+		/// </summary>
+		/// <param name="propertyName">Name of the property</param>
+		bool IsPropertyModified(string propertyName);
+	}
+
+	public class HookedEntity : IHookedEntity
+	{
+		private Type _entityType;
+
 		public HookedEntity(DbEntityEntry entry)
 		{
 			Entry = entry;
 			InitialState = (EntityState)entry.State;
 		}
 
-		/// <summary>
-		/// Gets the hooked entity entry
-		/// </summary>
 		public DbEntityEntry Entry
 		{
 			get;
@@ -24,18 +58,20 @@ namespace SmartStore.Core.Data.Hooks
 			get { return Entry.Entity as BaseEntity; }
 		}
 
-		/// <summary>
-		/// Gets the initial (presave) state of the hooked entity
-		/// </summary>
+		public Type EntityType
+		{
+			get
+			{
+				return _entityType ?? (_entityType = this.Entity?.GetUnproxiedType());
+			}
+		}
+
 		public EntityState InitialState
 		{
 			get;
-			internal set;
+			set;
 		}
 
-		/// <summary>
-		/// Gets or sets the current state of the hooked entity
-		/// </summary>
 		public EntityState State
 		{
 			get
@@ -48,9 +84,6 @@ namespace SmartStore.Core.Data.Hooks
 			}
 		}
 
-		/// <summary>
-		/// Gets a value indicating whether the entity state has changed during hooking.
-		/// </summary>
 		public bool HasStateChanged
 		{
 			get
@@ -59,10 +92,6 @@ namespace SmartStore.Core.Data.Hooks
 			}
 		}
 
-		/// <summary>
-		/// Gets a value indicating whether a property has been modified.
-		/// </summary>
-		/// <param name="propertyName">Name of the property</param>
 		public bool IsPropertyModified(string propertyName)
 		{
 			Guard.NotEmpty(propertyName, nameof(propertyName));