+++ src/Tests/SmartStore.Core.Tests/TopologicSortTests.cs
@@ -6,8 +6,7 @@ using System.Linq;
 using SmartStore.Collections;
 
 namespace SmartStore.Core.Tests
-{
-	
+{	
 	[TestFixture]
 	public class TopologicSortTests
     {